Fix Uncontrolled to Controlled error
When filling out the form, following error happened:
Warning: MyForm is changing an uncontrolled input of type text to be controlled. Input elements should not switch from uncontrolled to controlled (or vice versa). Decide between using a controlled or uncontrolled input element for the lifetime of the component

Turns out React doesn't like addidtional state to be added when user interacting with the form. Thus, setting all forms keys when init the comonent would solve the problem.